Notification LED: Fix blinking the wrong color for notification
Currently, the notification blinks red. This is not correct as on stock
it blinks green. This is because the system sends WHITE to the liblights
module. If green is sent, the LED will blink green.

This does not change the charging behavior. If charging LED is red, Green
will flash. If charging LED is green, red shall flash. This behavior is
to allow users to see a notification while charging

Change-Id: I7c83d41fb87ce616cd7192f72e770236e588abbd